requirements.txt
gradio>=4.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
matplotlib>=3.5.0
pillow>=9.0.0
tqdm>=4.64.0
fastapi>=0.68.0
uvicorn>=0.15.0
seaborn==0.13.2